requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
docx==0.2.4
docx2txt==0.8
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MouseInfo==0.1.3
mysql-connector-python==9.1.0
narwhals==1.8.2
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pillow==10.4.0
plotly==5.24.1
plotly-express==0.4.1
protobuf==5.28.2
psycopg==3.2.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==17.0.0
PyAutoGUI==0.9.54
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pymssql==2.3.1
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytweening==1.2.0
pytz==2024.2
pywin32==308
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
schedule==1.2.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
statsmodels==0.14.4
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.6
